Conditional Sleeping Carp Price
While Sleeping Carp is ridiculous, it's not quite Romerol levels of strength. There isn't a lot of reason for a traitor to be unable to buy it as they aren't just doing combat stuff (unlike IAA, who are in a deathmatch). However, IAA still needs to get a nerf to Sleeping Carp, as it's an autoinclude for a first purchase.